Added the RestDesktop client I wrote for the FuelAPI project.
It consumes the GET methods from the FuelAPI project in order to show
a plotted chart of the price changes of a particular fuel.

It uses JFreeChart to draw the diagrams, and also the build in Java httpClient and jackson serializer in order to
deserialize the json data received from the api's endpoints.